+++ packages/backend/src/controllers/savedChartController.ts
@@ -4,10 +4,8 @@ import {
     ApiGetChartHistoryResponse,
     ApiGetChartVersionResponse,
     ApiSuccessEmpty,
-    DashboardFilters,
     DateGranularity,
     SortField,
-    TimeFrames,
 } from '@lightdash/common';
 import {
     Body,
@@ -78,6 +76,7 @@ export class SavedChartController extends Controller {
             dashboardFilters: any; // DashboardFilters; temp disable validation
             invalidateCache?: boolean;
             dashboardSorts: SortField[];
+            dashboardUuid: string;
             granularity?: DateGranularity;
         },
         @Path() chartUuid: string,
@@ -93,6 +92,7 @@ export class SavedChartController extends Controller {
                 invalidateCache: body.invalidateCache,
                 dashboardSorts: body.dashboardSorts,
                 granularity: body.granularity,
+                dashboardUuid: body.dashboardUuid,
             }),
         };
     }